feat: fetchableContainer 기능 구현 #177
